pyproject.toml
[project]
name = "text2audio"
version = "0.1.0"
description = "A modular CLI utility for converting text/Markdown to spoken audio using OpenAI TTS"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.106.1",
    "python-dotenv>=1.1.1",
    "regex>=2025.9.1",
    "rich>=14.1.0",
]

[project.scripts]
text2audio = "text2audio.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]
